refactor(mcp): extract stdio transport
